Moved CourseType property from CourseRun classes to Course classes
