+++ src/Dfc.CourseDirectory.Core/DataStore/Sql/Queries/CreateCourseRun.cs
@@ -26,7 +26,6 @@ namespace Dfc.CourseDirectory.Core.DataStore.Sql.Queries
         public CourseAttendancePattern? AttendancePattern { get; set; }
         public CourseStudyMode? StudyMode { get; set; }
         public UserInfo CreatedBy { get; set; }
-        public DateTime CreatedOn { get; set; }
-        public CourseType? CourseType { get; set; }
+        public DateTime CreatedOn { get; set; }        
     }
 }
