+++ src/Dfc.CourseDirectory.Core/DataStore/Sql/QueryHandlers/CreateCourseHandler.cs
@@ -35,7 +35,8 @@ INSERT INTO Pttcd.Courses (
     HowYoullLearn,
     WhatYoullNeed,
     HowYoullBeAssessed,
-    WhereNext
+    WhereNext,
+    CourseType
 ) VALUES (
     @CourseId,
     {(int)CourseStatus.Live},
@@ -52,7 +53,8 @@ INSERT INTO Pttcd.Courses (
     @HowYoullLearn,
     @WhatYoullNeed,
     @HowYoullBeAssessed,
-    @WhereNext
+    @WhereNext,
+    @CourseType
 )
 
 INSERT INTO Pttcd.CourseRuns (
@@ -76,8 +78,7 @@ INSERT INTO Pttcd.CourseRuns (
     DurationValue,
     StudyMode,
     AttendancePattern,
-    [National],
-    CourseType
+    [National]    
 )
 SELECT
     CourseRunId,
@@ -100,8 +101,7 @@ SELECT
     DurationValue,
     StudyMode,
     AttendancePattern,
-    [National],
-    CourseType
+    [National]
 FROM @CourseRuns
 
 INSERT INTO Pttcd.CourseRunSubRegions (CourseRunId, RegionId)
@@ -127,6 +127,7 @@ EXEC Pttcd.RefreshFindACourseIndex @CourseRunIds = @CourseRunIds, @Now = @Create
                 query.WhatYoullNeed,
                 query.HowYoullBeAssessed,
                 query.WhereNext,
+                query.CourseType,
                 CourseRuns = TvpHelper.CreateCourseRunsTable(query.CourseRuns.Select(cr => (
                     cr.CourseRunId,
                     cr.CourseName,
@@ -142,8 +143,8 @@ EXEC Pttcd.RefreshFindACourseIndex @CourseRunIds = @CourseRunIds, @Now = @Create
                     cr.DurationValue,
                     cr.StudyMode,
                     cr.AttendancePattern,
-                    cr.National,
-                    cr.CourseType))),
+                    cr.National
+                    ))),
                 SubRegions = TvpHelper.CreateCourseRunSubRegionsTable(
                     from cr in query.CourseRuns
                     from subRegionId in cr.SubRegionIds ?? Array.Empty<string>()
