+++ src/Dfc.CourseDirectory.Core/DataStore/Sql/TvpHelper.cs
@@ -40,8 +40,8 @@ namespace Dfc.CourseDirectory.Core.DataStore.Sql
                 int DurationValue,
                 CourseStudyMode? StudyMode,
                 CourseAttendancePattern? AttendancePattern,
-                bool? National,
-                CourseType? CourseType)> rows)
+                bool? National
+                )> rows)
         {
             var table = new DataTable();
             table.Columns.Add("CourseRunId", typeof(Guid));
@@ -58,8 +58,7 @@ namespace Dfc.CourseDirectory.Core.DataStore.Sql
             table.Columns.Add("DurationValue", typeof(int));
             table.Columns.Add("StudyMode", typeof(byte));
             table.Columns.Add("AttendancePattern", typeof(byte));
-            table.Columns.Add("National", typeof(bool));
-            table.Columns.Add("CourseType", typeof(int));
+            table.Columns.Add("National", typeof(bool));            
 
             foreach (var row in rows)
             {
@@ -79,8 +78,7 @@ namespace Dfc.CourseDirectory.Core.DataStore.Sql
                     row.DurationValue,
                     row.StudyMode,
                     row.AttendancePattern,
-                    row.National,
-                    row.CourseType
+                    row.National                    
                 });
             }
