+++ src/Dfc.CourseDirectory.WebV2/Features/ChooseQualification/FlowModel.cs
@@ -45,7 +45,8 @@ namespace Dfc.CourseDirectory.WebV2.Features.ChooseQualification
           string howYouWillLearn,
           string whatYouWillNeedToBring,
           string howYouWillBeAssessed,
-          string whereNext)
+          string whereNext,
+          CourseType? courseType)
         {
             WhoThisCourseIsFor = whoThisCourseIsFor;
             EntryRequirements = entryRequirements;
@@ -54,6 +55,7 @@ namespace Dfc.CourseDirectory.WebV2.Features.ChooseQualification
             WhatYouWillNeedToBring = whatYouWillNeedToBring;
             HowYouWillBeAssessed = howYouWillBeAssessed;
             WhereNext = whereNext;
+            CourseType = courseType;
         }
 
         public void SetDeliveryMode(CourseDeliveryMode? deliveryMode)
@@ -74,8 +76,8 @@ namespace Dfc.CourseDirectory.WebV2.Features.ChooseQualification
         CourseDurationUnit? durationUnit,
         CourseStudyMode? studyMode,
         CourseAttendancePattern? attendancePattern,
-        Guid? venueId,
-        CourseType? courseType)
+        Guid? venueId
+        )
         {
             CourseName = courseName;
             ProviderCourseRef = providerCourseRef;
@@ -90,8 +92,7 @@ namespace Dfc.CourseDirectory.WebV2.Features.ChooseQualification
             DurationUnit = durationUnit;
             AttendancePattern = attendancePattern;
             StudyMode = studyMode;
-            VenueId = venueId;
-            CourseType = courseType;
+            VenueId = venueId;            
         }
     }
 }
